# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/appointment": {
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Handles the creation of a new appointment.",
                "operationId": "2157315882445629255bdd9b2461ce66",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "patient_id": {
                                        "description": "The user ID associated with the patient.",
                                        "type": "string",
                                        "example": "faeac40d-a9c0-4c4a-aa6b-ebb210afb589"
                                    },
                                    "nutritionist_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "12345"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "enum": [
                                            "catering",
                                            "nutritional_advice"
                                        ],
                                        "example": "catering"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "PatientModel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "string",
                                            "example": "12345"
                                        },
                                        "full_name": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john.doe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/v1/appointment/{id}": {
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Create a new appointment",
                "description": "Handles the creation of a new appointment.",
                "operationId": "16ad5d8e07a3694c82adce1e9e24d575",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del appointment",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "PatientModel created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "string",
                                            "example": "12345"
                                        },
                                        "full_name": {
                                            "properties": {
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Doe"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john.doe@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
...
    },
...
}